added canScrollVertically
Added method canScrollVertically as the equivalent of
canScrollHorizontally but for a VerticalViewPager implementation such as
the one provided by castorflex
(https://github.com/castorflex/VerticalViewPager).

As VerticalViewPager is written to use this method, it wasn't necessary
to add a canScrollVerticallyFroyo wrapper like what has been done for
the horizontal direction to support API v8.